+++ examples/demo-rollup/src/celestia_rollup.rs
@@ -10,7 +10,7 @@ use sov_mock_zkvm::{MockCodeCommitment, MockZkvm};
 use sov_modules_api::default_spec::{DefaultSpec, ZkDefaultSpec};
 use sov_modules_api::{CryptoSpec, Spec};
 use sov_modules_rollup_blueprint::{RollupBlueprint, WalletBlueprint};
-use sov_modules_stf_blueprint::StfBlueprint;
+use sov_modules_stf_blueprint::{RuntimeEndpoints, StfBlueprint};
 use sov_prover_storage_manager::ProverStorageManager;
 use sov_risc0_adapter::host::Risc0Host;
 use sov_rollup_interface::zk::aggregated_proof::CodeCommitment;
@@ -76,11 +76,10 @@ impl RollupBlueprint for CelestiaDemoRollup {
         sequencer_db: &SequencerDb,
         da_service: &Self::DaService,
         rollup_config: &RollupConfig<Self::DaConfig>,
-    ) -> Result<(jsonrpsee::RpcModule<()>, axum::Router<()>), anyhow::Error> {
+    ) -> anyhow::Result<RuntimeEndpoints> {
         let sequencer = rollup_config.da.own_celestia_address.clone();
 
-        #[allow(unused_mut)]
-        let (mut rpc_methods, axum_router) = sov_modules_rollup_blueprint::register_endpoints::<
+        let mut endpoints = sov_modules_rollup_blueprint::register_endpoints::<
             Self,
             ModAuth<Self::NativeSpec, Self::DaSpec>,
         >(
@@ -96,10 +95,10 @@ impl RollupBlueprint for CelestiaDemoRollup {
         crate::eth::register_ethereum::<Self::NativeSpec, Self::DaService>(
             da_service.clone(),
             storage.clone(),
-            &mut rpc_methods,
+            &mut endpoints.jsonrpsee_module,
         )?;
 
-        Ok((rpc_methods, axum_router))
+        Ok(endpoints)
     }
 
     async fn create_da_service(